main view fixes
