+++ client/src/components/PrinterRow.vue
@@ -1,41 +1,20 @@
 <script setup lang="ts">
-import { jobTime, useAssignComment, useGetFile, useGetJobFile, useReleaseJob, useRemoveJob, useStartJob, type Job } from '@/model/jobs';
+import { useGetJobFile, useReleaseJob, useStartJob, type Job } from '@/model/jobs';
 import { printers, useSetStatus, type Device } from '@/model/ports';
-import { type Issue, useGetIssues, useAssignIssue } from '../model/issues'
 import { useRouter } from 'vue-router';
-import GCode3DImageViewer from '@/components/GCode3DImageViewer.vue'
-import GCode3DLiveViewer from '@/components/GCode3DLiveViewer.vue';
-import { onMounted, ref } from 'vue';
 
 const props = defineProps<{
     printer: Device;
+    openModal: (job: Job, printerName: string, num: number, printer: Device) => Promise<void>;
+    setJob: (job: Job) => Promise<void>;
 }>();
 
 const { setStatus } = useSetStatus();
 const { start } = useStartJob()
 const { releaseJob } = useReleaseJob()
-const { removeJob } = useRemoveJob()
-const { getFile } = useGetFile()
-const { issues } = useGetIssues()
-const { assign } = useAssignIssue()
-const { assignComment } = useAssignComment()
 const { getFileDownload } = useGetJobFile()
 
 const router = useRouter();
-let currentJob = ref<Job>();
-let currentPrinter = ref<Device>();
-const selectedJob = ref<Job>()
-const selectedIssue = ref<Issue>()
-let issuelist = ref<Array<Issue>>([])
-
-let isGcodeImageVisible = ref(false)
-let isGcodeLiveViewVisible = ref(false)
-let jobComments = ref('')
-
-onMounted(async () => {
-    const retrieveissues = await issues()
-    issuelist.value = retrieveissues
-})
 
 const sendToQueueView = (printer: Device | undefined) => {
     if (printer) {
@@ -59,28 +38,7 @@ const startPrint = async (printerid: number, jobid: number) => {
     await start(jobid, printerid)
 }
 
-const setJob = async (job: Job) => {
-    jobComments.value = job.comment || '';
-    selectedJob.value = job;
-}
 
-const openModal = async (job: Job, printerName: string, num: number, printer: Device) => {
-    await jobTime(job, printers)
-    currentJob.value = job
-    currentJob.value.printer = printerName
-    currentPrinter.value = printer
-    if (num == 1) {
-        isGcodeLiveViewVisible.value = true
-    } else if (num == 2) {
-        isGcodeImageVisible.value = true
-        if (currentJob.value) {
-            const file = await getFile(currentJob.value)
-            if (file) {
-                currentJob.value.file = file
-            }
-        }
-    }
-}
 
 const openPrinterInfo = (printer: Device) => {
     printer.isInfoExpanded = !printer.isInfoExpanded;
@@ -89,157 +47,9 @@ const openPrinterInfo = (printer: Device) => {
 const releasePrinter = async (jobToFind: Job | undefined, key: number, printerIdToPrintTo: number) => {
     await releaseJob(jobToFind, key, printerIdToPrintTo)
 }
-
-function formatTime(milliseconds: number): string {
-    const seconds = Math.floor((milliseconds / 1000) % 60)
-    const minutes = Math.floor((milliseconds / (1000 * 60)) % 60)
-    const hours = Math.floor((milliseconds / (1000 * 60 * 60)) % 24)
-
-    const hoursStr = hours < 10 ? '0' + hours : hours
-    const minutesStr = minutes < 10 ? '0' + minutes : minutes
-    const secondsStr = seconds < 10 ? '0' + seconds : seconds
-
-    if ((hoursStr + ':' + minutesStr + ':' + secondsStr === 'NaN:NaN:NaN')) return '<i>idle</i>'
-    return hoursStr + ':' + minutesStr + ':' + secondsStr
-}
-
-function formatETA(milliseconds: number): string {
-    const date = new Date(milliseconds)
-    const timeString = date.toLocaleTimeString('en-US', { hour: '2-digit', minute: '2-digit', second: '2-digit', hour12: true })
-
-
-    if (isNaN(date.getTime()) || timeString === "07:00 PM") {
-        return '<i>idle</i>'
-    }
-
-    return date.toLocaleTimeString('en-US', { hour: '2-digit', minute: '2-digit', hour12: true })
-}
-
-const doAssignIssue = async () => {
-    if (selectedJob.value === undefined) return
-    await releasePrinter(selectedJob.value, 3, selectedJob.value.printerid)
-
-    if (selectedIssue.value !== undefined) {
-        await assign(selectedIssue.value.id, selectedJob.value.id)
-    }
-    await assignComment(selectedJob.value, jobComments.value)
-    selectedJob.value.comment = jobComments.value
-    selectedIssue.value = undefined
-    selectedJob.value = undefined
-}
-
 </script>
 
 <template>
-    <div class="modal fade" id="issueModal" tabindex="-1" aria-labelledby="assignIssueLabel" aria-hidden="true"
-        data-bs-backdrop="static">
-        <div class="modal-dialog modal-dialog-centered">
-            <div class="modal-content">
-                <div class="modal-header d-flex align-items-end">
-                    <h5 class="modal-title mb-0" id="assignIssueLabel" style="line-height: 1;">Job #{{ selectedJob?.id
-                        }}</h5>
-                    <h6 class="modal-title" id="assignIssueLabel" style="padding-left:10px; line-height: 1;">{{
-                        selectedJob?.date }}</h6>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"
-                        @click="selectedIssue = undefined; selectedJob = undefined;"></button>
-                </div>
-                <div class="modal-body">
-                    <form @submit.prevent="">
-                        <div class="mb-3">
-                            <label for="issue" class="form-label">Select Issue</label>
-                            <select name="issue" id="issue" v-model="selectedIssue" class="form-select" required>
-                                <option disabled value="undefined">Select Issue</option>
-                                <option v-for="issue in issuelist" :value="issue">
-                                    {{ issue.issue }}
-                                </option>
-                            </select>
-                        </div>
-                    </form>
-                    <div class="form-group mt-3">
-                        <label for="exampleFormControlTextarea1">Comments</label>
-                        <textarea class="form-control" id="exampleFormControlTextarea1" rows="3"
-                            v-model="jobComments"></textarea>
-                    </div>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal"
-                        @click="selectedIssue = undefined; selectedJob = undefined">Close</button>
-                    <button type="button" class="btn btn-success" data-bs-dismiss="modal" @click="doAssignIssue">Save
-                        Changes</button>
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <!-- bootstrap 'gcodeLiveViewModal' -->
-    <div class="modal fade" id="gcodeLiveViewModal" tabindex="-1" aria-labelledby="gcodeLiveViewModalLaebl"
-        aria-hidden="true" @shown.bs.modal="isGcodeLiveViewVisible = true"
-        @hidden.bs.modal="isGcodeLiveViewVisible = false">
-        <div class="modal-dialog modal-dialog-centered modal-xl">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="gcodeLiveViewModalLabel">
-                        <b>{{ currentJob?.printer }}:</b> {{ currentJob?.name }}
-                    </h5>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-                </div>
-                <div class="modal-body">
-                    <div class="row">
-                        <GCode3DLiveViewer v-if="isGcodeLiveViewVisible" :job="currentJob" />
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <!-- bootstrap 'gcodeImageModal' -->
-    <div class="modal fade" id="gcodeImageModal" tabindex="-1" aria-labelledby="gcodeImageModalLabel" aria-hidden="true"
-        @shown.bs.modal="isGcodeImageVisible = true" @hidden.bs.modal="isGcodeImageVisible = false">
-        <div class="modal-dialog modal-dialog-centered modal-xl">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="gcodeImageModalLabel">
-                        <b>{{ currentJob?.printer }}:</b> {{ currentJob?.name }}
-                    </h5>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-                </div>
-                <div class="modal-body">
-                    <div class="row">
-                        <GCode3DImageViewer v-if="isGcodeImageVisible" :job="currentJob" />
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <!-- bootstrap 'gcodeModal' -->
-    <div class="modal fade" id="gcodeModal" tabindex="-1" aria-labelledby="gcodeModalLabel" aria-hidden="true">
-        <div class="modal-dialog modal-dialog-centered modal-xl">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="gcodeModalLabel"><b>{{ currentJob?.printer }}:</b> {{ currentJob?.name
-                        }}</h5>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-                </div>
-                <div class="modal-body">
-                    <div class=" row">
-                        <div class="col-sm-12">
-                            <div class="card bg-light mb-3">
-                                <div class="card-body">
-                                    <h5 class="card-title">
-                                        <pre> .GCODE VIEWER </pre>
-                                    </h5>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-
-
-
     <td
         v-if="(printer.status == 'printing' || printer.status == 'complete' || printer.status == 'paused' || printer.status == 'colorchange' || (printer.status == 'offline' && (printer.queue?.[0]?.status == 'complete' || printer.queue?.[0]?.status == 'cancelled')))">
         {{ printer.queue?.[0].id }}
@@ -373,7 +183,7 @@ const doAssignIssue = async () => {
                     </div>
                 </div>
                 <div type="button" class="btn btn-danger" data-bs-toggle="modal" data-bs-target="#issueModal"
-                    @click=setJob(printer.queue!![0])>
+                    @click=setJob(printer.queue![0])>
                     Fail
                 </div>
             </div>
@@ -394,9 +204,9 @@ const doAssignIssue = async () => {
 
     </td>
 
-    <td style="width: 1%; white-space: nowrap; height: 55px; padding-top: 5px;">
+    <td style="width: 1%; white-space: nowrap;">
         <div style="display: flex; justify-content: space-between; align-items: center;">
-            <i class="fa" :class="['fa-chevron-down', printer.isInfoExpanded ? 'rotate-up' : 'rotate-down']"
+            <i class="fa fa-chevron-down" :class="printer.isInfoExpanded ? 'rotate-down' : 'rotate-up'"
                 @click="openPrinterInfo(printer)"></i>
             <div :class="{ 'not-draggable': printer.queue && printer.queue.length == 0 }" class="dropdown">
                 <div style="display: flex; justify-content: center; align-items: center; height: 100%;">
@@ -430,12 +240,41 @@ const doAssignIssue = async () => {
     </td>
 
     <td class="text-center handle" :class="{ 'not-draggable': printers.length <= 1 || printer.isInfoExpanded }"
-        :rowspan="printer.isInfoExpanded ? 2 : 1" :style="{ 'vertical-align': printer.isInfoExpanded ? 'middle' : '' }">
-        <i class="fas fa-grip-vertical" :class="{ 'icon-disabled': printers.length <= 1 || printer.isInfoExpanded }"></i>
+        :rowspan="printer.isInfoExpanded ? 3 : 1" :style="{ 'vertical-align': printer.isInfoExpanded ? 'middle' : '' }">
+        <i class="fas fa-grip-vertical"
+            :class="{ 'icon-disabled': printers.length <= 1 || printer.isInfoExpanded }"></i>
     </td>
 </template>
 
 <style scoped>
+@keyframes rotateDown {
+    0% {
+        transform: rotate(0deg);
+    }
+
+    100% {
+        transform: rotate(180deg);
+    }
+}
+
+@keyframes rotateUp {
+    0% {
+        transform: rotate(180deg);
+    }
+
+    100% {
+        transform: rotate(0deg);
+    }
+}
+
+.rotate-down {
+    animation: rotateDown 0.3s ease-in-out forwards;
+}
+
+.rotate-up {
+    animation: rotateUp 0.3s ease-in-out forwards;
+}
+
 .border-extended {
     position: relative;
 }
@@ -449,18 +288,6 @@ const doAssignIssue = async () => {
     width: 1px;
     background: #929292;
     height: calc(100% + 1.5px);
-}
-
-.fa {
-    transition: transform 0.3s ease-in-out;
-}
-
-.rotate-up {
-    transform: rotate(0deg);
-}
-
-.rotate-down {
-    transform: rotate(180deg);
 }
 
 .truncate {
@@ -479,6 +306,10 @@ const doAssignIssue = async () => {
     flex: 1 0 auto;
     margin: 0 0.375rem;
     flex-shrink: 0;
+}
+
+td {
+    height: 32px !important;
 }
 
 .dropdown-item {
